Animations with no frameSync should be synchronized to the start time
(i.e. "start time" means the time when running was set to true)

Unfortunately, we reset the timestamp *after* the image had been
loaded and assembled in buildNode(), therefore, the first *displayed*
frame would always be the first one in the animation regardless of
how long time it took to load the image.

Change-Id: I6b57a0f69afd3c9140628b874481cdeda1dffce5
Reviewed-by: Alan Alpert (Personal) <416365416c@gmail.com>