Do not fail test because window moves out of desktop area bounds.
For some reason the window moved in the bottom-right direction each
time the window was resized. This could cause the window to move out of
the bounds of the desktop area, which again could cause an autotest
failure. We therefore ensure that the window is centered for each test
call.

Task-number: QTBUG-33017
Change-Id: I0a2b9bb4ba866d8a3f80309c6b33863bab4d2bcf
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@digia.com>