Do not crash under certain conditions in AnimatedSprite
The code that tried to find the remaning number of frames on the last
row was incorrect if the last row were full.
It basically did totalNumberOfRows % framesPerRow, which doesn't always
work, for instance:

If the animation had 8 frames per row and a total of 16 frames
the calculation to find the remaining number of frames became:

 16 % 8 = 0 (correct result should have been 8)

As you can see, it only didn't work in the cases where the last row
were full.

Change-Id: I9ff6ac7baa91538b01576b659ea6651992cfca0b
Reviewed-by: Michael Brasser <michael.brasser@live.com>