Only delete all child nodes twice if effectRefEffectivelyChanged is set.
We already handle ChildrenUpdateMask again in the general case below, so this
second clear doesn't make sense, *but* for the fact that we are modifying
groupNode here when effectRefEffectivelyChanged is set. So keep doing it for
that case.

As a nice side-effect, this means that groupNode is not always allocated.

Also, don't fetch groupNode redundantly if we are clearing nodes.

Task-number: QTBUG-34255
Change-Id: Icc01b942d3c09c17300e19d61e9a65f220004796
Reviewed-by: Gunnar Sletta <gunnar@sletta.org>