QQuickParticleSystem: Optimize hash manipulation.
Before even checking the hash, perform the fast (empty string) check.

If it's not empty, then do a single hash lookup, not two of them. There's no
point doing a second lookup, since if it doesn't exist, we insert anyway.

This takes hash manipulation on samegame on a large screen from ~142ms to ~70ms,
which is expected.

Change-Id: I4e2a099e996f2f5fb95a5b066d7756373a1dbacd
Reviewed-by: Gunnar Sletta <gunnar@sletta.org>