Relayout rich text when width increases
In updateSize() we were trying to force relayouts for RichText
when the width changes by always setting widthExceeded to true.
But further down in the same function, we overwrote this
with textWidth() < idealWidth(), which doesn't work, because
both properties are the wrapped width of the document and
should only differ if the text cannot be wrapped properly.

The result was that when increasing the width of a Text element,
we would hit the optimization and skip the relayout.

[ChangeLog][Text] Fixed Text with the RichText format to
correctly update wrapping when the width of the element grows.

Change-Id: I5fd87052a5ba7e8ee2549be0cfac4adc8ddf8290
Task-number: QTBUG-33020
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>