Fix syntax error when trying to declare read-only object properties
The grammar did not allow for the declaration of

    readonly property QtObject foo: QtObject { ... }

and it required a workaround through an alias:

    readonly property alias foo: _foo
    property QtObject _foo: QtObject { ... }

This was merely a glitch in the grammar, I see no reason not to support this.
The semantics are like a const pointer in C++, the property itself is read-only
but the object pointed to has per-property defined read/write semantics.

Task-number: QTBUG-41971
Change-Id: I99e2e7ed58731e387a38e46ec39922d280a21ceb
Reviewed-by: Michael Brasser <michael.brasser@live.com>
Reviewed-by: Lars Knoll <lars.knoll@digia.com>