Simplify QFont value type wrapper in preparation for gadgets
The value type wrapper for QFont tries to keep track whether pixel or point
sizes were set and warns the programmer accordingly if both were set. This is
implemented using separate boolean variables, but this can also be done using
QFont's resolve mask.

This move is motivated by the ability of the value type wrappers to operate
only on the type they wrap in the future, without requiring additional
helper variables.

Change-Id: Ia05093b5dcc950281ef89b2bb6fe96c18329cdfa
Reviewed-by: Lars Knoll <lars.knoll@digia.com>
Reviewed-by: Michael Brasser <michael.brasser@live.com>